requirements.txt
aenum==3.1.16
annotated-types==0.7.0
arabic-reshaper==3.0.0
asgiref==3.11.0
asn1crypto==1.5.1
cashfree_pg==5.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
cloudinary==1.44.1
cryptography==46.0.5
cssselect2==0.9.0
dj-database-url==3.1.2
Django==5.2.9
django-cloudinary-storage==0.3.0
freetype-py==2.5.1
gunicorn==21.2.0
html5lib==1.1
idna==3.11
lxml==6.0.2
oscrypto==1.3.0
packaging==26.0
pillow==12.1.0
psycopg2-binary==2.9.11
pycairo==1.29.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyHanko==0.33.0
pyhanko-certvalidator==0.29.1
pypdf==6.7.5
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
razorpay==2.0.0
reportlab==4.4.10
requests==2.32.5
rlPyCairo==0.4.0
sentry-sdk==1.32.0
six==1.17.0
sqlparse==0.5.5
svglib==1.6.0
tinycss2==1.5.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritools==6.0.1
urllib3==2.0.7
webencodings==0.5.1
whitenoise==6.12.0
xhtml2pdf==0.2.17
